// src/types/socios.ts
export interface Socio {
	id: string;
	nome: string;
	nacionalidade: string;
	estadoCivil: number;
	maioridadeCivil: string;
	profissao: string;
	rg: string;
	cpf: string;
	endereco: string;
	nroImovel: string;
	complemento: string;
	bairro: string;
	cidade: string;
	uf: string;
	cep: string;
	userEmail: string;
}